+++ file.json
@@ -1,0 +1,1 @@
+{"Amenity.de283621-6479-47b7-99de-a412da36086c": {"id": "de283621-6479-47b7-99de-a412da36086c", "created_at": "2023-11-30T07:46:14.790209", "updated_at": "2023-11-30T07:46:14.790209", "__class__": "Amenity"}, "BaseModel.db42552f-2509-417d-8389-38d66a6e4217": {"id": "db42552f-2509-417d-8389-38d66a6e4217", "created_at": "2023-11-30T07:46:14.790216", "updated_at": "2023-11-30T07:46:14.790216", "__class__": "BaseModel"}, "City.a225c752-5be3-4c50-bf4d-e34f2a7f9750": {"id": "a225c752-5be3-4c50-bf4d-e34f2a7f9750", "created_at": "2023-11-30T07:46:14.790223", "updated_at": "2023-11-30T07:46:14.790223", "__class__": "City"}, "Place.91ad174a-b5af-4660-9258-f59c4df03f9a": {"id": "91ad174a-b5af-4660-9258-f59c4df03f9a", "created_at": "2023-11-30T07:46:14.790229", "updated_at": "2023-11-30T07:46:14.790229", "__class__": "Place"}, "Review.4878d883-2c83-466a-9575-2621833980b5": {"id": "4878d883-2c83-466a-9575-2621833980b5", "created_at": "2023-11-30T07:46:14.790234", "updated_at": "2023-11-30T07:46:14.790234", "__class__": "Review"}, "State.933a551b-54c2-4b5c-93fa-b529399ef09a": {"id": "933a551b-54c2-4b5c-93fa-b529399ef09a", "created_at": "2023-11-30T07:46:14.790239", "updated_at": "2023-11-30T07:46:14.790239", "__class__": "State"}, "User.04fd8fd0-6b05-4151-ab86-a2920193ed13": {"id": "04fd8fd0-6b05-4151-ab86-a2920193ed13", "created_at": "2023-11-30T07:46:14.790246", "updated_at": "2023-11-30T07:46:14.790246", "__class__": "User"}}
